refactor: 提取 Anthropic 为独立 provider, 增加 InferFromBaseURL
将 Anthropic Messages API 相关代码从 OpenAIProvider 中抽取到独立的
AnthropicProvider, 消除 OpenAI provider 中的 anthropic 分支逻辑。

- 新增 AnthropicProvider: 请求/响应序列化, 流式解析, 认证头
- provider.Resolve 通过 InferFromBaseURL 从 BaseURL 自动推断 provider
- NewProviderFromResolved 根据 provider 名路由到对应实现
- 重命名 wrapOpenAIReadError → wrapReadError (两端共用)
